requirements.txt
streamlit
opencv-python-headless
face-recognition
numpy
pandas
Pillow

